modern_scada/platform/core/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
asyncpg>=0.28.0
aiosqlite>=0.19.0
pymodbus>=3.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
apscheduler>=3.10.0
pyyaml>=6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
websockets>=11.0
python-multipart>=0.0.6
psutil>=5.9.0
pytest
httpx
pytest-asyncio
redis>=5.0.0
